Fixed MetricLogger after restart train from checkpoint

+++ loggers/MetricLogger.py
@@ -1,3 +1,4 @@
+import os
 import numpy as np
 import time, datetime
 import matplotlib.pyplot as plt
@@ -6,13 +7,14 @@ import matplotlib.pyplot as plt
 class MetricLogger:
     def __init__(self, save_dir):
         self.save_log = None if save_dir is None else save_dir / "log"
-        if self.save_log is not None:
-            with open(self.save_log, "w") as f:
-                f.write(
-                    f"{'Episode':>8}{'Step':>8}{'Epsilon':>10}{'MeanReward':>15}"
-                    f"{'MeanLength':>15}{'MeanLoss':>15}{'MeanQValue':>15}"
-                    f"{'TimeDelta':>15}{'Time':>20}\n"
-                )
+        if self.save_log is not None: 
+            if not os.path.exists(self.save_log):
+                with open(self.save_log, "w") as f:
+                    f.write(
+                        f"{'Episode':>8}{'Step':>8}{'Epsilon':>10}{'MeanReward':>15}"
+                        f"{'MeanLength':>15}{'MeanLoss':>15}{'MeanQValue':>15}"
+                        f"{'TimeDelta':>15}{'Time':>20}\n"
+                    )
             self.ep_rewards_plot = save_dir / "reward_plot.jpg"
             self.ep_lengths_plot = save_dir / "length_plot.jpg"
             self.ep_avg_losses_plot = save_dir / "loss_plot.jpg"
